services/delivery-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
prometheus-client
requests